backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.135.1",
    "langchain>=1.2.10",
    "langchain-openai>=1.1.10",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
